+++ CHANGELOG.md
@@ -34,6 +34,7 @@
     - nicer output when processing from STDIN
     - aggregation more human readable
     - exclude filter + filter menu (composed from unique, exclude and include filter)
+    - remove some of the processing settings
 * internal
     - daemon pipes stderr as well
     - got rid of python-Levenshtein package and did not use psutil package so installing whole gcc is not needed anymore